Profile: properly handle recursion in print flat format
Previously, we'd double-count any recursive function for each repetition
of that function in a stack snapshot.
Now we properly account for the time spent in any particular
frame by only recording a single count for it per stack snapshot,
regardless of how often that frame appeared in that stacktrace
due to recursion (and independent of inlining).